// src/store/modules/types/user.ts
export interface SignUpInfo {
  id: Nullable<string>;
  consNo: Nullable<string>;
  consName: Nullable<string>;
  accountId: Nullable<string>;
  provinceCode: Nullable<string>;
  orgNo: Nullable<string>;
  houseType: Nullable<string>;
  livingArea: Nullable<string>;
  peopleNumber: Nullable<string>;
  equipment: Nullable<string>;
  deviceDescription: Nullable<string>;
  createTime: Nullable<string>;
  updateTime: Nullable<string>;
  phoneNo: Nullable<string>;
}